+++ syslog.php
@@ -1184,11 +1184,11 @@ function syslog_filter($sql_where, $tab) {
 	?>
 	<script type='text/javascript'>
 	initSyslogMain({
-		pageTab: '<?php print get_request_var('tab'); ?>',
-		placeHolder: '<?php print __esc('Enter a search term', 'syslog'); ?>',
-		noneSelectedText: '<?php print __esc('Select Device(s)', 'syslog'); ?>',
-		devicesSelectedText: '<?php print __esc('Devices Selected', 'syslog'); ?>',
-		allDevicesText: '<?php print __esc('All Devices Selected', 'syslog'); ?>'
+		pageTab: <?php print syslog_json_safe(get_request_var('tab')); ?>,
+		placeHolder: <?php print syslog_json_safe(__('Enter a search term', 'syslog')); ?>,
+		noneSelectedText: <?php print syslog_json_safe(__('Select Device(s)', 'syslog')); ?>,
+		devicesSelectedText: <?php print syslog_json_safe(__('Devices Selected', 'syslog')); ?>,
+		allDevicesText: <?php print syslog_json_safe(__('All Devices Selected', 'syslog')); ?>
 	});
 	</script>
 	<?php